+++ crates/intrinsic-test/src/arm/json_parser.rs
@@ -2,7 +2,7 @@ use super::constraint::Constraint;
 use super::intrinsic::ArmIntrinsicType;
 use crate::common::argument::{Argument, ArgumentList};
 use crate::common::intrinsic::Intrinsic;
-use crate::common::intrinsic_types::{IntrinsicType, IntrinsicTypeDefinition};
+use crate::common::intrinsic_helpers::{IntrinsicType, IntrinsicTypeDefinition};
 use serde::Deserialize;
 use serde_json::Value;
 use std::collections::HashMap;
